Have `spine` subpackage for internal dependency objects
