+++ api/build.gradle.kts
@@ -26,9 +26,11 @@
 
 import io.spine.internal.dependency.Aedile
 import io.spine.internal.dependency.Jackson
-import io.spine.internal.dependency.Spine
-import io.spine.internal.dependency.ToolBase
-import io.spine.internal.dependency.Validation
+import io.spine.internal.dependency.spine.CoreJava
+import io.spine.internal.dependency.spine.Logging
+import io.spine.internal.dependency.spine.Spine
+import io.spine.internal.dependency.spine.ToolBase
+import io.spine.internal.dependency.spine.Validation
 
 plugins {
     `build-proto-model`
@@ -39,14 +41,14 @@ dependencies {
     listOf(
         Spine.base,
         Spine.text,
-        Spine.CoreJava.server,
+        CoreJava.server,
         ToolBase.lib,
         ToolBase.psiJava
     ).forEach {
         api(it)
     }
 
-    api(Spine.Logging.lib)
+    api(Logging.lib)
 
     implementation(Spine.reflect)
     implementation(Aedile.lib)
