+++ build.gradle.kts
@@ -28,7 +28,8 @@
 
 import io.spine.internal.dependency.Dokka
 import io.spine.internal.dependency.Grpc
-import io.spine.internal.dependency.Spine
+import io.spine.internal.dependency.spine.CoreJava
+import io.spine.internal.dependency.spine.Spine
 import io.spine.internal.gradle.RunBuild
 import io.spine.internal.gradle.publish.PublishingRepos
 import io.spine.internal.gradle.publish.SpinePublishing
@@ -41,7 +42,7 @@ import org.jetbrains.dokka.gradle.DokkaMultiModuleTask
 
 buildscript {
     standardSpineSdkRepositories()
-    val baseForBuildScript = io.spine.internal.dependency.Spine.baseForBuildScript
+    val baseForBuildScript = io.spine.internal.dependency.spine.Spine.baseForBuildScript
     dependencies {
         classpath(io.spine.internal.dependency.Protobuf.GradlePlugin.lib)
         classpath(baseForBuildScript)
@@ -94,7 +95,7 @@ allprojects {
                 Spine.reflect,
                 Spine.base,
                 Spine.testlib,
-                Spine.server
+                CoreJava.server
             )
         }
     }
