+++ buildSrc/src/main/kotlin/build-proto-model.gradle.kts
@@ -1,11 +1,11 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -27,8 +27,8 @@
 @file:Suppress("RemoveRedundantQualifierName")
 
 import io.spine.internal.dependency.Protobuf
-import io.spine.internal.dependency.McJava
-import io.spine.internal.dependency.Spine
+import io.spine.internal.dependency.spine.McJava
+import io.spine.internal.dependency.spine.Spine
 
 /**
  * The dependency onto Spine Validation causes the circular dependency in this Gradle project.
