+++ buildSrc/src/main/kotlin/io/spine/internal/dependency/spine/ProtoData.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.internal.dependency
+package io.spine.internal.dependency.spine
 
 /**
  * Dependencies on ProtoData modules.
