+++ gradle-api/build.gradle.kts
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,12 +24,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.Spine
+import io.spine.internal.dependency.spine.ToolBase
 
 dependencies {
     compileOnly(gradleApi())
-    implementation(Spine.toolBase)
+    implementation(ToolBase.lib)
     implementation(project(":api"))
     testImplementation(gradleApi())
-    testImplementation(Spine.pluginTestlib)
+    testImplementation(ToolBase.pluginTestlib)
 }
